0.2.12
Added support for interpolating the T1 image to show the region where the pca was measured, as well as any 3D mask that is registered to the T1.

Added a static module: SELMAInterpolate with all of the auxillary functions.
SELMAT1Dicom contains all the logic for managing the interpolation.

The framecounter now only displays 1 frame when the T1 is shown.

The following bugs still exist:
-The interpolated result is shifted in the L-R direction, unknown why.
-The method only seems to work for dicoms where the frames are in correct order. TODO: add sorting
-The method currently only supports sagittal T1s. Coronal and Axial views are necessary for testing