0.3.0
Added support for segmentation via SPM.

SPM12 and dcm2nii need to be installed on the computer.
If no paths to these are found, the program prompts the user for the path.

The t1 image needs to be in a separate folder for the algorithm to work.

A matlab script is launched that creates a .nii file from the t1 dicom. That .nii file is then used to make 6-tissue segmentations with spm. The white matter segmentation is returned, interpolated to the right slice and thresholded to yield the new mask.

(still buggy)