0.3.1
Small changes to make sure the results are comparable to Lennart's old version.

Zero crossings are now calculated with the corrected velocity frames instead of the uncorrected ones.

min and max velocity are now based on the absolute values of the velocities.

renamed brainmaskprob to whitematterprob in the settings.

Pycharm removed some formatting in SELMAT1Dicom, it's ugly now :'(

fixed small bug in spmSegment.m