Disabled clustering of significant flow with isointense magnitudes
Clusters of voxels with significant flow with isointense magnitudes were erroneously detected as vessels. This led to an unrealistic high amount of detected vessels in 3T and 7T data. Furthermore, actual vessels that were duplicates of these isointense clusters were rejected, which led to the misdetection of actual vessels.

For now, the lines that clustered these isointense voxels are commented.