SELMA V1.1
In this new update of SELMA, the algorithm of the tool was validated against the MATLAB implementation and several inconsistencies were fixed/removed. The following major bugs were fixed:
- Fixed increased amount of detected isointense clusters in SELMA compared to MATLAB. New clustering algorithm is implemented which is completely identical to MATLAB. The new algorithm makes use of the newly implemented 'remove_ccs_from_mask' function which is analogous to the original MATLAB implementation. This functions checks whether voxels identified with either a positive or negative significant magnitude are belonging to a vessel and subsequently labels the entire vessel as positive or negative and removes them from the mask. The end result is that less voxels with significant flow and isointense magnitude remain to be identified.
- Fixed inconsistency in the remove non-perpendicular step. The previous SELMA implementation was not similar enough to the original MATLAB implementation. The new implementation using the 'regionprops' function removes a lot of the inconsistencies, but it is still not exactly the same due to function differences between Python and MATLAB. Edge cases might exist where MATLAB includes vessels based on axes ratio that SELMA might exclude. This might be fixed in future versions if deemed important enough.

After the necessary fixes, SELMA produced results that were nearly identical to the MATLAB implementation using the RESOLUTION dataset. This gives us enough confidence that SELMA is ready to be used on a larger scale for different studies and datasets.

Additional to the bug fixes, new features were built into SELMA. Now the tool is able to compute the average velocity and normalised PI per dicom scan. SELMA outputs this in a new text file which also keeps track of the amount of (non-)perpendicular, cluster or lone vessels it found. Furthermore, SELMA keeps track of the amount of positive, negative, and isointense clusters it found for full transparency to how many vessels SELMA found at each step.