1.2.0
From version 1.2.0 and onwards, batch analysis is SELMA is supported. Please check the README for further documentation on how to properly utilize the new batch analysis function. The following changes are made in this version to accommodate for the new batch analysis function:
- Updated README with new information on batch analysis.
- Updated main GUI to enable warning message prior to batch analysis.
- Added functionality to write batch analysis results for all dicom files in a single Python dictionary.
- Added functionality to save Python dictionaries during batch analysis as a .mat file.

Additionally, several changes are made to accommodate a future update where a variable threshold for significant flow will be introduced. These changes are currently not active, but the code is already present.