1.3.0
From version 1.3.0 and onwards, batch analysis on classic DICOM files in SELMA is supported. Please check the README for further documentation on how to properly utilize the new batch analysis function for classic DICOM files. The following changes are made in this version to accommodate for the new batch analysis function:

- Updated README with new information on batch analysis on classic DICOM files.

This version of SELMA also adds a new structure tab to the settings. In this structure tab, a specific anatomical structure can be selected for analysis. This automatically enables the correct cluster settings for vessel detection. For advanced users, there is also a setting which can override the pre-applied cluster settings of the selected structure. By navigating to the Advanced Clustering tab in the settings, the user can freely select which magnitudes and flows to include in the analysis.

This version also fixes a rare bug where mislabeled magnitude or phase frames were omitted in the analysis, resulting in a dimension mismatch. This has been fixed by performing a label check before grouping all scan frames.