1.3.1
In this version of SELMA the algorithm for deduplication was slightly changed to better match the original iMblob implementation in the PULSATE study. Previously, SELMA would throw out both vessels if they were within 6 voxels of each other. Now SELMA first checks whether those vessels belong to the same magnitude blob and will only keep the vessel with the highest velocity for every magnitude blob.

Furthermore, a small tweak to the remove non-perpendicular step was added to the code. This tweak is meant to make this step more visually intuitive. Previously, the shape of only the magnitude blob was considered for perpendicularity. Now blobs containing voxels with significant flow and magnitude above the perpendicularity threshold are assessed for perpendicularity. Approval for this change in the algorithm is pending.

Lastly, a failsafe was built in which catches any attempt to run SELMA without prespecified structures or clusters. The user will be prompted to select the correct structure or clusters in the settings menu.

Bug fixes:
- Fixed a bug where sometimes the incorrect amount of vessels would be displayed in the output.