SELMA 1.4.0: Final preparations for v2
In this update of SELMA we have made the final preparations in the code for an eventual v2 update to SELMA. In this large version update of SELMA we have changed the algorithm that computes the noise in the velocity maps. Now we estimate this noise using the velocity distribution, rather than using the signal to noise ratio in the magnitude. Further documentation on this change will follow in the 2.0 update.

Furthermore, we have made some changes to the deduplicate step to make it less strict and also take the respective magnitude blobs into account.

Finally, we have added the filename to the batch analysis results to make further analysis easier. Bug fixes were also made to the batch analysis code.