SELMA 1.4.9
Rudimentary fix for GE scans has been applied to this version of SELMA. Now, in all folders containing GE scans should be accompanied with a Scan_Parameters.txt file which contains the necessary information to load in the scans. This file contains three lines which represent the heart rate, Turbo Field Echo, and Repetition Time, respectively.  This text file can be used to overrule the parameters found in the DICOM headers. Thus, these parameters could also be changed with user input in case SELMA cannot find the proper values in the DICOM headers (the case with GE scanners) An empty line means that for the corresponding parameter the value from the DICOM header will be chosen. This text file has to be added to every folder with GE scans otherwise the analysis will not work.