SELMA 2.0
Final push to finalise SELMA 2.0 after quality testing the pipeline. In this push, a safety mechanism was built in for rare instances where enhanced DICOM files would be reconstructed in wrong order, thus leading to bad velocity traces.

This version will be used for analysis of the ZOOM@Rotterdam data.